FileWatcher now resembles an actual app.  it doesn't do anything on the GUI (yet), but it stops bouncing. Also includes use of NSAppleScript for scripting, a nib-instantiated object that starts the old main() in a thread, and everything is run within the bundle.  This does mean it only is a matter of installing the app, and i believe it doesn't care where it is put.
FIXME: the PurgeList is in the plugins of the app bundle.  if the app is replaced, thats gone.  should be in ~/Library/Application\ Support/FileWatcher (or the like).  for now, plugins are actually stil easy to get to: Finder, get into, plugin tab lets you add it, remove it, or install a new one.  That might be something to add a GUI for.